requirements.txt
python-telegram-bot[job-queue]>=21.0,<22
aiohttp>=3.9,<4
aiosqlite>=0.20,<1
python-dotenv>=1.0,<2
